default to local file transport, share code with plugin
