+++ core/api/src/initializers/files.ts
@@ -1,16 +1,19 @@
 import { api, Initializer } from "actionhero";
 import { File } from "../models/File";
+import { string } from "prop-types";
 
 export class Files extends Initializer {
   constructor() {
     super();
     this.name = "files";
     this.loadPriority = 999;
+    this.startPriority = 1;
   }
 
   async initialize() {
     api.files = {
       types: ["csv", "json", "image", "video", "audio", "other"],
+      transport: undefined,
       downloadToServer: async (file: File) => {
         throw new Error("not implemented");
       },
@@ -21,5 +24,24 @@ export class Files extends Initializer {
         throw new Error("not implemented");
       },
     };
+  }
+
+  async start() {
+    // default to localFile if no file transport was loaded by a plugin
+    if (!api.files.transport) {
+      const { FileTransportLocal } = require("../classes/fileTransportLocal");
+      const instance = new FileTransportLocal();
+      api.files.downloadToServer = async (file) => {
+        return instance.downloadToServer(file);
+      };
+
+      api.files.set = async (type, remotePath, localFile) => {
+        return instance.set(type, remotePath, localFile);
+      };
+
+      api.files.destroy = async (file) => {
+        return instance.destroy(file);
+      };
+    }
   }
 }
