+++ plugins/@grouparoo/files-local/README.md
@@ -4,4 +4,4 @@ Store Grouparoo files on your local hard drive.
 
 ## Configuration
 
-Set the environment variable `FILES_LOCAL_STORAGE` to set your storage folder (default: `<project>/files`)
+Set the environment variable `FILES_LOCAL_STORAGE` to set your storage folder (default: `<CWD>/files-${process.env.NODE_ENV}`). For example, `./files-development`
