+++ plugins/@grouparoo/files-s3/src/initializers/fileTransportS3.ts
@@ -16,6 +16,7 @@ export class EmailAuth extends Initializer {
 
     const instance = new FileTransportS3();
 
+    api.files.transport = packageJSON.name;
     api.files.downloadToServer = async (file) => {
       return instance.downloadToServer(file);
     };
